requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
requests
imageio-ffmpeg
quran-transcript
gspread
google-auth
python-telegram-bot

